+++ src/Common/src/Interop/Linux/libc/Interop.epoll_wait.cs
@@ -3,15 +3,12 @@
 
 using System;
 using System.Runtime.InteropServices;
-using System.Text;
 
 internal static partial class Interop
 {
     internal static partial class libc
     {
-        public const int FIONREAD = 0x541b;
-
         [DllImport(Libraries.Libc, SetLastError = true)]
-        public static extern unsafe int ioctl(int d, UIntPtr request, void* argp);
+        public static extern unsafe int epoll_wait(int epfd, epoll_event* events, int maxevents, int timeout);
     }
 }
